+++ src/frontend/src/lib/components/tournaments/registration/SoloSquadTournamentRegister.svelte
